icons moved form assets to components
