+++ src/components/icons/IconCheckboxChecked.jsx
